layers: add missing tracking for vk_khr_display
When rendering directly to the display the validation
layers do not record the creation of a VkDisplay object
in the ThreadSafety class. This causes a failed assertion
when the VkDisplay is next seen, rendering validation
unusable.

This commit is only to demonstrate what the resulting
changes should look like, it does not add these to the
autogeneration process. This needs to be added to
the generation scripts before it's ready to be committed.